agents/scraper/requirements.txt
crewai>=0.28.0
langchain>=0.1.1
langchain-openai>=0.0.3
aiohttp>=3.8.6
python-dotenv>=1.0.0
pydantic>=2.4.0
requests>=2.31.0
pytest>=7.4.0
lxml>=4.9.3
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
tenacity>=8.2.3
